weeks/week08/projectone/requirements.txt
rpi-lgpio
smbus2
gradio
